fix ctrl
