--- controllers/auth-controllers.js
@@ -1,106 +1,0 @@
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const gravatar = require("gravatar");
-const fs = require("fs/promises");
-const path = require("path");
-const Jimp = require("jimp");
-
-const { ctrlWrapper } = require("../utils");
-
-const { HttpError } = require("../helpers");
-
-const { User } = require("../models/user");
-
-const { SECRET_KEY } = process.env;
-
-const avatarsDir = path.join(__dirname, "../", "public", "avatars");
-
-const register = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (user) {
-    throw HttpError(409, "Email already in use");
-  }
-  const hashPassword = await bcrypt.hash(password, 10);
-  console.log(email);
-  const avatarURL = gravatar.url(email, { d: "identicon" });
-  console.log(avatarURL);
-  const result = await User.create({
-    ...req.body,
-    password: hashPassword,
-    avatarURL,
-  });
-  res.status(201).json({
-    name: result.name,
-    email: result.email,
-  });
-};
-
-const login = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (!user) {
-    throw HttpError(401, "Email or password1 is wrong");
-  }
-  const passwordCompare = await bcrypt.compare(password, user.password);
-  if (!passwordCompare) {
-    throw HttpError(401, "Email or password2 is wrong");
-  }
-  const payload = {
-    id: user._id,
-  };
-  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
-  await User.findByIdAndUpdate(user._id, { token });
-  res.json({ token });
-};
-
-const getCurrent = async (req, res) => {
-  const { name, email } = req.user;
-  res.json({
-    name,
-    email,
-  });
-};
-
-const logout = async (req, res) => {
-  const { _id } = req.user;
-  await User.findByIdAndUpdate(_id, { token: "" });
-  res.json({ message: "Logout success" });
-};
-
-// const updateStatusSubscription = async (req, res) => {
-//   const  { _id } = req.user;
-//   const result = await User.findByIdAndUpdate(_id, req.body, {
-//     new: true,
-//   });
-//   if (!result) {
-//     throw HttpError(404, `Contact with ${contactId} not found`);
-//   }
-//   res.json(result);
-// };
-
-const updateAvatar = async (req, res) => {
-  const { _id } = req.user;
-  const { path: tmpUpload, filename } = req.file;
-  const avatarName = `${_id}_${filename}`;
-
-  const resultUpload = path.join(avatarsDir, avatarName);
-  Jimp.read(resultUpload, (err, lenna) => {
-    if (err) throw err;
-    lenna.resize(250, 250).quality(60).greyscale().write(resultUpload);
-  });
-  await fs.rename(tmpUpload, resultUpload);
-  const avatarURL = path.join("avatars", avatarName);
-  await User.findByIdAndUpdate(_id, { avatarURL });
-
-  res.json({ avatarURL });
-};
-
-module.exports = {
-  register: ctrlWrapper(register),
-  login: ctrlWrapper(login),
-  getCurrent: ctrlWrapper(getCurrent),
-  logout: ctrlWrapper(logout),
-  // updateStatusSubscription: ctrlWrapper(updateStatusSubscription),
-  updateAvatar: ctrlWrapper(updateAvatar),
-};
