+++ controllers/auth/getCurrent.js
@@ -1,0 +1,10 @@
+const getCurrent = async (req, res) => {
+  const { name, email } = req.user;
+  res.json({
+    name,
+    email,
+  });
+};
+
+
+module.exports = getCurrent;
