+++ controllers/auth/logout.js
@@ -1,0 +1,9 @@
+const { User } = require("../../models/user");
+
+const logout = async (req, res) => {
+  const { _id } = req.user;
+  await User.findByIdAndUpdate(_id, { token: "" });
+  res.json({ message: "Logout success" });
+};
+module.exports = logout;
+
