--- controllers/contacts-controllers.js
@@ -1,80 +1,0 @@
-const { ctrlWrapper } = require("../utils");
-
-const { HttpError } = require("../helpers");
-
-const { Contact } = require("../models/contact");
-
-const listContacts = async (req, res) => {
-  const { _id: owner } = req.user;
-  const { page = 1, limit = 10, favorite } = req.query;
-  const skip = (page - 1) * limit;
-  const result = await Contact.find({ owner }, "-createdAt, -updatedAt", {
-    skip,
-    limit,
-  }).populate("owner", "name email");
-  if (!result) {
-    res.status(400);
-    throw new Error("Controller: Unnable to fetch contacts");
-  }
-  if (!favorite) {
-    return res.json(result);
-  } else {
-    const favoriteContact = result.filter((contact) => {if(contact.favorite.toString() === favorite){return contact}});
-    return res.json(favoriteContact);
-  }
-};
-
-const getContactById = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
-  if (!result) {
-    throw HttpError(404, `Contact with ${contactId} not found`);
-  }
-  res.json(result);
-};
-
-const addContact = async (req, res) => {
-  const { _id: owner } = req.user;
-  const result = await Contact.create({ ...req.body, owner });
-  res.status(201).json(result);
-};
-
-const removeContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndDelete(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json({ message: "contact deleted" });
-};
-
-const updateContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `Contact with ${contactId} not found`);
-  }
-  res.json(result);
-};
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `Contact with ${contactId} not found`);
-  }
-  res.json(result);
-};
-
-module.exports = {
-  listContacts: ctrlWrapper(listContacts),
-  getContactById: ctrlWrapper(getContactById),
-  addContact: ctrlWrapper(addContact),
-  removeContact: ctrlWrapper(removeContact),
-  updateContact: ctrlWrapper(updateContact),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-};
