+++ controllers/contacts/addContact.js
@@ -1,0 +1,9 @@
+const { Contact } = require("../../models");
+
+const addContact = async (req, res) => {
+  const { _id: owner } = req.user;
+  const result = await Contact.create({ ...req.body, owner });
+  res.status(201).json(result);
+};
+
+module.exports = addContact
