+++ controllers/contacts/removeContact.js
@@ -1,0 +1,13 @@
+const { HttpError } = require("../../helpers");
+
+const { Contact } = require("../../models");
+
+const removeContact = async (req, res) => {
+    const { contactId } = req.params;
+    const result = await Contact.findByIdAndDelete(contactId);
+    if (!result) {
+      throw HttpError(404, "Not found");
+    }
+    res.json({ message: "contact deleted" });
+  };
+  module.exports = removeContact
